pyproject.toml
[project]
name = "fluidgenie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Peiyan Yang", email = "yangpeiyan1992@gmail.com" }]
dependencies = [
    "einops>=0.8.2",
    "flax>=0.10.7",
    "imageio>=2.37.2",
    "jax>=0.6.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "optax>=0.2.6",
    "orbax-checkpoint>=0.11.0",
    "phiflow>=3.4.0",
    "pillow>=12.1.0",
    "tensorboardx>=2.6.4",
    "tqdm>=4.67.2",
    "tyro>=0.9.3",
]

[project.optional-dependencies]
cuda12 = [
    "jax[cuda12]>=0.6.2; platform_system == 'Linux'",
]
cuda13 = [
    "jax[cuda13]>=0.9.0; platform_system == 'Linux' and python_version >= '3.11'",
]
